// v2/public.d.ts
export interface PublicKey extends Key {
    readonly algorithm: {
        readonly name: 'PASETO v2.public';
    };
    readonly type: 'public';
    readonly version: 2;
    readonly kind: 'public';
}
// ...
export interface SecretKey extends Key {
    readonly algorithm: {
        readonly name: 'PASETO v2.public';
    };
    readonly type: 'secret';
    readonly version: 2;
    readonly kind: 'secret';
}
// ...
export declare const ImportPublicKeyFactory: ReturnType<typeof PublicImportPublicKey<2, PublicKey>>;
/**
 * Built-in PASERK k2.public key-export capability factory.
 *
 * @category Key Management
 */
export declare const ExportPublicKeyFactory: ReturnType<typeof PublicExportPublicKey<2, PublicKey>>;
/**
 * Built-in PASERK k2.secret key-import capability factory.
 *
 * @category Key Management
 */
export declare const ImportSecretKeyFactory: ReturnType<typeof PublicImportSecretKey<2, SecretKey>>;